(update) move full input to utils
