+++ client/components/Utils/FullInput.jsx
@@ -1,0 +1,13 @@
+import styled from 'react-emotion';
+
+const FullInput = styled('input')`
+  width: 100%;
+  height: 100%;
+  position: absolute;
+  top: 0;
+  left: 0;
+  opacity: 0;
+  cursor: pointer;
+`;
+
+export default FullInput;
